requirements.txt
# Flask and extensions
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1

# AWS
boto3==1.34.34

# Firebase
firebase-admin==6.4.0

# Gemini AI
google-generativeai==0.3.2

# Image processing
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80

# Environment and configuration
python-dotenv==1.0.1

# Testing
pytest==8.0.1

# HTTP and API
requests==2.31.0

# Excel handling
openpyxl==3.1.2

# JWT handling
python-jose==3.3.0

# Email validation
email-validator==2.1.0.post1

# CORS handling
Flask-CORS==4.0.0

# Rate limiting
Flask-Limiter==3.5.0

# Caching
Flask-Caching==2.1.0

# Session handling
Flask-Session==0.6.0

# Background tasks
celery==5.3.6
redis==5.0.1

# Logging
python-json-logger==2.0.7

# Security
bcrypt==4.1.2
cryptography==42.0.2

# Database
google-cloud-firestore==2.14.0

# Date and time handling
pytz==2024.1

# Form handling
WTForms==3.1.2

# Development tools
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0